backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet==3.2.4

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP 客户端（用于微信API、DeepSeek API）
httpx==0.26.0

# 安全认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 日期时间处理
python-dateutil==2.8.2

# 环境变量
python-dotenv==1.0.0

# AI/RAG 相关
chromadb==0.4.22
sentence-transformers==2.3.1
rank-bm25==0.2.2
jieba==0.42.1

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
